requirement.txt
asgiref==3.6.0
bleach==5.0.1
blinker==1.6.3
boto==2.49.0
boto3==1.24.22
botocore==1.27.22
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==41.0.0
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
dj-database-url==0.5.0
Django==4.1.7
django-allauth==0.51.0
django-ckeditor==6.4.2
django-crispy-forms==2.0
django-environ==0.10.0
django-extra-keywords==1.1.2
django-gunicorn==0.1.1
django-import-export==3.1.0
django-js-asset==2.0.0
django-plotly-dash==2.2.2
django-quill-editor==0.1.40
django-storages==1.12.3
django-summernote==0.8.20.0
django-user-agents==0.4.0
djangorestframework==3.14.0
dpd-components==0.1.0
et-xmlfile==1.1.0
Flask==3.0.0
fontawesomefree==6.3.0
fonttools==4.37.4
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.3
MarkupPy==1.14
MarkupSafe==2.1.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==9.4.0
plotly==5.18.0
pycparser==2.21
PyJWT==2.4.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
tenacity==8.2.3
typing_extensions==4.6.2
tzdata==2023.3
ua-parser==0.18.0
urllib3==1.26.16
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==3.0.1
xlrd==2.0.1
xlwt==1.3.0
